djangoblog/contact/forms.py
from django import forms
from .models import Contact
from captcha.fields import CaptchaField


class ContactForm(forms.ModelForm):
    captcha = CaptchaField()

    class Meta:
        model = Contact
        fields = ('name', 'email', 'text', 'captcha')